requirements.txt
fastapi
uvicorn[standard]
scikit-learn
pandas
numpy
joblib
python-dotenv
matplotlib
seaborn